Deploying openyurt on top of kubernetes & knative unit test with github workflows change template to yaml
