+++ scripts/openyurt-deployer/cloudNodePoolTemplate.yaml
@@ -1,0 +1,6 @@
+apiVersion: apps.openyurt.io/v1beta1
+kind: NodePool
+metadata:
+  name: poolName
+spec:
+  type: Cloud
