--- scripts/openyurt-deployer/configs/go.mod
@@ -1,3 +1,0 @@
-module github.com/vhive-serverless/vhive/scripts/openyurt_deployer/configs
-
-go 1.20
