--- scripts/openyurt-deployer/configs/vhive.go
@@ -1,19 +1,0 @@
-package configs
-
-type VHiveConfigStruct struct {
-	FirecrackerKernelImgDownloadUrl string
-	StargzVersion                   string
-	VHiveRepoPath                   string
-	VHiveRepoBranch                 string
-	VHiveRepoUrl                    string
-	VHiveSetupConfigPath            string
-	ForceRemote                     bool
-}
-
-var VHive = VHiveConfigStruct{
-	VHiveRepoPath:        ".",
-	VHiveRepoBranch:      "main",
-	VHiveRepoUrl:         "https://github.com/vhive-serverless/vHive.git",
-	VHiveSetupConfigPath: "../../configs/setup",
-	ForceRemote:          false,
-}
