+++ scripts/openyurt-deployer/go.mod
@@ -1,15 +1,18 @@
 module github.com/vhive-serverless/vhive/scripts/openyurt_deployer
 
-go 1.20
+go 1.19
+
+require (
+	github.com/sfreiberg/simplessh v0.0.0-20220719182921-185eafd40485
+	github.com/sirupsen/logrus v1.9.3
+	github.com/vhive-serverless/vHive/scripts/utils v0.0.0-20231114022852-400a49d284cb
+)
 
 require (
 	github.com/davidmz/go-pageant v1.0.2 // indirect
 	github.com/kr/fs v0.1.0 // indirect
 	github.com/pkg/sftp v1.13.4 // indirect
-	github.com/sfreiberg/simplessh v0.0.0-20220719182921-185eafd40485 // indirect
-	github.com/sirupsen/logrus v1.9.3 // indirect
-	github.com/vhive-serverless/vHive/scripts/configs v0.0.0-20231018022901-6a0c478d2c9f // indirect
-	github.com/vhive-serverless/vHive/scripts/utils v0.0.0-20231018022901-6a0c478d2c9f // indirect
+	github.com/vhive-serverless/vHive/scripts/configs v0.0.0-20231114022852-400a49d284cb // indirect
 	golang.org/x/crypto v0.0.0-20220411220226-7b82a4e95df4 // indirect
 	golang.org/x/sys v0.0.0-20220715151400-c0bba94af5f8 // indirect
 )
