--- scripts/openyurt-deployer/node/go.mod
@@ -1,3 +1,0 @@
-module github.com/vhive-serverless/vhive/scripts/openyurt_deployer/node
-
-go 1.20
