--- scripts/openyurt-deployer/template/go.mod
@@ -1,3 +1,0 @@
-module github.com/vhive-serverless/vhive/scripts/openyurt_deployer/template
-
-go 1.20
