--- scripts/openyurt-deployer/template/template.go
@@ -1,3 +1,0 @@
-package template
-
-var vHiveConfigsURL = "https://raw.githubusercontent.com/anshalshukla/vHive/release-1.9/configs"
