+++ scripts/openyurt-deployer/yurtTemplate.yaml
@@ -1,0 +1,63 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  labels:
+    k8s-app: yurt-hub
+  name: yurt-hub
+  namespace: kube-system
+spec:
+  volumes:
+  - name: hub-dir
+    hostPath:
+      path: /var/lib/yurthub
+      type: DirectoryOrCreate
+  - name: kubernetes
+    hostPath:
+      path: /etc/kubernetes
+      type: Directory
+  - name: pem-dir
+    hostPath:
+      path: /var/lib/kubelet/pki
+      type: Directory
+  containers:
+  - name: yurt-hub
+    image: openyurt/yurthub:latest
+    imagePullPolicy: IfNotPresent
+    volumeMounts:
+    - name: hub-dir
+      mountPath: /var/lib/yurthub
+    - name: kubernetes
+      mountPath: /etc/kubernetes
+    - name: pem-dir
+      mountPath: /var/lib/kubelet/pki
+    command:
+    - yurthub
+    - --v=2
+    - --server-addr=https://__kubernetes_master_address__
+    - --node-name=$(NODE_NAME)
+    - --join-token=__bootstrap_token__
+    livenessProbe:
+      httpGet:
+        host: 127.0.0.1
+        path: /v1/healthz
+        port: 10267
+      initialDelaySeconds: 300
+      periodSeconds: 5
+      failureThreshold: 3
+    resources:
+      requests:
+        cpu: 150m
+        memory: 150Mi
+      limits:
+        memory: 300Mi
+    securityContext:
+      capabilities:
+        add: ["NET_ADMIN", "NET_RAW"]
+    env:
+    - name: NODE_NAME
+      valueFrom:
+        fieldRef:
+          fieldPath: spec.nodeName
+  hostNetwork: true
+  priorityClassName: system-node-critical
+  priority: 2000001000
